perf: Refractor and memoize ZoneHeaderTitle.tsx
